Add multiple labels prediction test, remove integration tests
